feature(Tasks): Switch planned and spent time from minutes to hours

+++ App/Commands/Tasks/Common/TaskOptions.cs
@@ -18,25 +18,25 @@ public static class TaskOptions
 
    public static Option<bool?> GetClosedOption()
    {
-         var value = new Option<bool?>(
-            name: "--closed",
-            getDefaultValue: () => null,
-            description: "Task closed");
-         value.AddAlias("-c");
-         return value;
+      var value = new Option<bool?>(
+         name: "--closed",
+         getDefaultValue: () => null,
+         description: "Task closed");
+      value.AddAlias("-c");
+      return value;
    }
 
    public static Option<int?> GetProjectIdOption()
    {
-         var value = new Option<int?>(
-            name: "--project-id",
-            getDefaultValue: () => null,
-            description: "Project id");
-         value.AddAlias("-p");
-         return value;
+      var value = new Option<int?>(
+         name: "--project-id",
+         getDefaultValue: () => null,
+         description: "Project id");
+      value.AddAlias("-p");
+      return value;
    }
 
-   public static Option<int?> GetPlannedTimeOption()
+   public static Option<decimal?> GetPlannedTimeOption()
    {
       return new(
          name: "--planned-time",
